gymbuddy-api/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.3.1
dj-database-url==2.2.0
firebase-admin==6.6.0
psycopg[binary]==3.2.3
gunicorn==23.0.0
whitenoise==6.8.2
sqlparse==0.5.5
